composeApp/src/commonTest/kotlin/br/com/noartcode/theprice/ui/presentation/home/HomeViewModelTest.kt
package br.com.noartcode.theprice.ui.presentation.home

import app.cash.turbine.test
import br.com.noartcode.theprice.ui.di.commonTestModule
import br.com.noartcode.theprice.ui.di.platformTestModule
import kotlinx.coroutines.Dispatchers
import kotlinx.coroutines.ExperimentalCoroutinesApi
import kotlinx.coroutines.test.UnconfinedTestDispatcher
import kotlinx.coroutines.test.resetMain
import kotlinx.coroutines.test.runTest
import kotlinx.coroutines.test.setMain
import org.koin.core.context.startKoin
import org.koin.core.context.stopKoin
import org.koin.test.KoinTest
import org.koin.test.inject
import kotlin.test.AfterTest
import kotlin.test.BeforeTest
import kotlin.test.Test
import kotlin.test.assertEquals

@OptIn(ExperimentalCoroutinesApi::class)
class HomeViewModelTest : KoinTest {

    private val viewModel:HomeViewModel by inject()

    @BeforeTest
    fun before() {
        Dispatchers.setMain(UnconfinedTestDispatcher())
        startKoin{ modules(commonTestModule(),  platformTestModule()) }
    }

    @AfterTest
    fun after() {
        Dispatchers.resetMain()
        stopKoin()
    }


    @Test
    fun `When the ViewModel started it should return an empty state object`() = runTest {

        // GIVEN
        viewModel.uiState.test {

            // WHEN
            val initialUiState = awaitItem()

            //THEN
            with(initialUiState) {
                assertEquals(emptyList(), payments)
                assertEquals(false, loading)
                assertEquals(null, errorMessage)
            }

            ensureAllEventsConsumed()
        }
    }

}